Improve support for Hugo shortcodes with no closing tag
The current `TokenIgnores` and `BlockIgnores` regexps do not fully
support shortcodes that do not contain a "closing tag", i.e. a
shortcode such as:

    {{< instagram BWNjjyYFxVx >}}

which appears in the official Hugo shortcode documentation:

https://gohugo.io/content-management/shortcodes/#instagram

I'm guessing that the new regexp's I added to `TokenIgnores` and
`BlockIgnores` are not 100% correct, however if you removed them and
run the unit test in `main_test.go` you'll see that Vale fails to
ignore the `myshortcode` shortcode on lines 56 and 60 and complains
about the words `myshortcode` and `Francia`.  I tested this with the
Vale v2.27.0.